update file metadata to 0.3.0b;
