+++ plugins/File Metadata/extractors/__init__.py
@@ -1,0 +1,3 @@
+from . import eze
+from . import ehentaidownloader
+from . import hdoujin
