+++ plugins/File Metadata/hplugin.json
@@ -2,7 +2,7 @@
     "id": "e38e24e4-8ca8-420e-b52b-c75510097653",
     "shortname": "file-metadata",
     "name": "File Metadata",
-    "version": "0.2.0b",
+    "version": "0.3.0b",
     "description": "Extracts and applies metadata from a file accompanying a gallery. Supports files produced from eze, e-hentai-downloader and hdoujin",
     "author": "Twiddly",
     "website": "https://github.com/happypandax/plugins/tree/master/plugins/File%20Metadata",
