+++ plugins/File Metadata/main.py
@@ -2,8 +2,9 @@ import __hpx__ as hpx
 import os
 import arrow
 import datetime
-import common
+import html
 import extractors
+from extractors import common
 
 log = hpx.get_logger(__name__)
 
@@ -37,49 +38,77 @@ def get_common_data(datatypes, fpath):
 language_model = hpx.command.GetModelClass("Language")
 title_model = hpx.command.GetModelClass("Title")
 artist_model = hpx.command.GetModelClass("Artist")
+parody_model = hpx.command.GetModelClass("Parody")
 circle_model = hpx.command.GetModelClass("Circle")
 url_model = hpx.command.GetModelClass("Url")
 namespacetags_model = hpx.command.GetModelClass("NamespaceTags")
 
 def apply_metadata(data, gallery):
+    """
+    data = {
+        'titles': None, # [(title, language),...]
+        'artists': None, # [(artist, (circle, circle, ..)),...]
+        'parodies': None, # [parody, ...]
+        'category': None,
+        'tags': None, # [tag, tag, tag, ..] or {ns:[tag, tag, tag, ...]}
+        'pub_date': None, # DateTime object or Arrow object
+        'language': None,
+        'urls': None # [url, ...]
+    }
+    """
+
     applied = False
 
-    log.debug("data:")
-    log.debug(f"{data}")
+    log.debug(f"data: {data}")
 
-    if isinstance(data['titles'], (list, tuple)):
+    if isinstance(data.get('titles'), (list, tuple)):
         for t, l in data['titles']:
             gtitle = None
             if t:
+                t = html.unescape(t)
                 gtitle = title_model(name=t)
             if t and l:
                 gtitle.language = language_model.as_unique(name=l)
             if gtitle:
                 gallery.update("titles", gtitle)
+        log.debug("applied titles")
         applied = True
 
-    if isinstance(data['artists'], (list, tuple)):
+    if isinstance(data.get('artists'), (list, tuple)):
         for a, c in data['artists']:
             if a:
-                gartist = artist_model.as_unique(name=a)
+                gartist = artist_model.as_unique(name=common.capitalize_text(a))
                 if not gartist in gallery.artists:
                     gallery.update("artists", gartist)
             if a and c:
                 for circlename in [x for x in c if x]:
-                    gcircle = circle_model.as_unique(name=circlename)
+                    gcircle = circle_model.as_unique(name=common.capitalize_text(circlename))
                     if not gcircle in gartist.circles:
                         gartist.update("circles", gcircle)
+        log.debug("applied artists")
         applied = True
 
-    if data['category']:
+    if isinstance(data.get('parodies'), (list, tuple)):
+        for p in data['parodies']:
+            if p:
+                gparody = parody_model.as_unique(name=common.capitalize_text(p))
+                if not gparody in gallery.parodies:
+                    gallery.update("parodies", gparody)
+
+        log.debug("applied parodies")
+        applied = True
+
+    if data.get('category'):
         gallery.update("category", name=data['category'])
+        log.debug("applied category")
         applied = True
     
-    if data['language']:
+    if data.get('language'):
         gallery.update("language", name=data['language'])
+        log.debug("applied language")
         applied = True
 
-    if isinstance(data['tags'], (dict, list)):
+    if isinstance(data.get('tags'), (dict, list)):
         if isinstance(data['tags'], list):
             data['tags'] = {None: data['tags']}
         ns_tags = []
@@ -95,22 +124,31 @@ def apply_metadata(data, gallery):
         for nstag in ns_tags:
             if nstag not in gallery.tags:
                 gallery.tags.append(nstag)
+        log.debug("applied tags")
         applied = True
 
-    if isinstance(data['pub_date'], (datetime.datetime, arrow.Arrow)):
+    if isinstance(data.get('pub_date'), (datetime.datetime, arrow.Arrow)):
         pub_date = data['pub_date']
         gallery.update("pub_date", pub_date)
+        log.debug("applied pub_date")
         applied = True
 
-    if isinstance(data['urls'], (list, tuple)):
+    if isinstance(data.get('urls'), (list, tuple)):
+        existing_urls = [x.name.lower() for x in gallery.urls]
         urls = []
         for u in data['urls']:
-            urls.append(url_model(name=u))
+            if u.lower() not in existing_urls:
+                urls.append(url_model(name=u))
         gallery.update("urls", urls)
+        log.debug("applied urls")
         applied = True
 
     return applied
     
+@hpx.subscribe("init")
+def inited():
+    common.plugin_config.update(hpx.get_plugin_config())
+
 @hpx.attach("GalleryFS.parse_metadata_file")
 def parse(path, gallery):
     fs = hpx.command.CoreFS(path)
